requirements.txt
# Project dependencies
rasa==3.6.14
streamlit==1.32.2
nltk==3.8.1
reportlab==4.0.9
scikit-learn==1.4.1.post1
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.31.0
flake8==7.0.0